Add more unit tests along with describe statements
This request includes the addition of more unit tests that explore more of the functionality of the erc20 token contract including the events that are emitted and different error messages. It also includes the addition of describe statements, which better organize the information within the tests. This makes it easier for people that are using the course to follow along with the tests in the GitHub because they will have a better understanding of what each section is for. It also more closely follows the manner in which testing was done in the course.