Update hardhat.config REPORT_GAS
If we set `REPORT_GAS` as `process.env.REPORT_GAS`, in case we have it defined in our `.env` we will always get `true` value from it no matter what it says. To see this we have to parse it to boolean and simply print it by `consolge.log`. We can fix this to work properly adding my code, so our `.env` can work good with "true" or "false" statements. Otherwise this `REPORT_GAS` will always create gas-report.txt if it is defined in .env